Changed the structure of the code in main so that it can be followed more easily **unfinished code in main**
